+++ spec/dummy/config/grouped-personas/dave.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+DemoMode.add_persona do
+  features << 'foo'
+  sign_in_as { DummyUser.create!(name: 'Dave') }
+end